refactor: simplify internal state, use AppDaemon state instead of HASS
Instead of tracking a list of hub IDs and a list of hub detail objects
as separate attributes, just make discovered_hubs a dictionary of
unique IDs where the values are the HubDetails objects.

Then after publishing a hub's discovery message, serialize those hub
objects in `discovered_attributes` and set that to AppDaemon state.

This way we can try to load that state during initialization.